Last attempt
Tried to fix the homing laser couldn't do it. tried to add collision for the player but that negates the shooting on the player so reverted that could not make another AI was too stressed after trying to add in the rest just have to hand it in not much help i was getting with a 2 person job for the amount of time.